requirements.txt
psycopg2-binary
protobuf
grpcio-tools
python-dotenv